Server/requirements.txt
websockets
pillow
flask
qrcode
pyOpenSSL
requests
python-dotenv
aiohttp
